Change directory structure
Without a sophisticated CUE package manager, users are left
to devise their own manual methods of managing and distributing modules.
We have settled on the use of git submodules which avoid checking in large
worktrees into repositories and offers some mechanism for upgrades, distribution,
and historic tracking.

CUE expects that dependencies exist in the cue.mod/pkg directory, but does not
make many other demands. Previously, this repo was structured so that it could get
dropped into a CUE project and take over the cue.mod. This is a problematic strategy
for a variety of reasons, namely, it prevents a CUE project having multiple CUE dependencies.

This commit addresses that shortcoming by restructuring the repository into a form that can fit
within the cue.mod/pkg directory of another project, namespaced under the module name. This module
as well as other modules dependent on this one, can all use the consistent module name for imports.
This is contrasted by an earlier strategy which treated modules paths as insignificant, leading to
import errors and time spent writing "massage" code.