enable publishing of the CUE module
I ran the following sequence of commands:

	# Add a language.version field
	cue mod fix

	# Set source to "self" and language version to a version later than v0.9
	cue mod edit --source self --language-version v0.11.0

	# Check that everything is ok
	cue vet ./...

Then manually fix up a bad reference to "byte" and a couple of unused
imports.

To publish this to the current module location
(`envoyproxy.io/envoy-cue`), there will need to be a file served from
`http://envoyproxy.io/.well-known/cue-central-registry.json` with the
following contents, or similar.

```
{
	auth: {
		type: "github"
		repo: "greymatter-io/envoy-cue"
	}
	cacheDuration: "24h"
}
```

Once that's in place, the module can be published to the Central Registry with:

```
cue mod publish v0.0.0
```

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>